+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 
 1. [](#new)
     * Added `Grav\Framework\Uri` classes extending PSR-7 `HTTP message UriInterface` implementation
+    * Added `$grav['uri]->getCurrentUri()` function to get `Grav\Framework\Uri` instance for the current URL.
     * Added ability to have `php` version dependencies in GPM assets
 1. [](#bugfix) 
-    * Fixed issue with remote PHP version determination for Grav updates    
+    * Fixed issue with remote PHP version determination for Grav updates
 
 # v1.4.0-rc.2
 ## 02/15/2018